test: refactor test files and add test for volar language plugins
